requirements.txt
aiogram==3.18.0
aiosqlite==0.20.0
python-dotenv==1.0.0
aiohttp==3.11.14
